pyproject.toml
[tool.poetry]
name = "satunnaisia-dissonansseja"
version = "0.1.0"
description = ""
authors = ["qm210 <quantenmace@gmail.com>", "nr4 <nr4@z10.info>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.1"
flask-migrate = "^4.0.5"
flask-socketio = "^5.3.6"
flask-sqlalchemy = "^3.1.1"
dependency-injector = "^4.41.0"
cached-path = "^1.5.1"
jinja2 = "^3.1.3"
psutil = "^5.9.8"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
